Unarchive a list of compressed tsv files into a database

Unarchive a list of compressed tsv files into a database

unark(files, db_con, streamable_table = NULL, lines = 50000L,
  overwrite = "ask", encoding = Sys.getenv("encoding", "UTF-8"),
  tablenames = NULL, ...)

Arguments

files
vector of filenames to be read in. Must be tsv format, optionally compressed using bzip2, gzip, zip, or xz format at present.

db_con
a database src (src_dbi object from dplyr)

streamable_table
interface for serializing/deserializing in chunks

lines
number of lines to read in a chunk.

overwrite
should any existing text files of the same name be overwritten? default is "ask", which will ask for confirmation in an interactive session, and overwrite in a non-interactive script. TRUE will always overwrite, FALSE will always skip such tables.

encoding
encoding to be assumed for input files.

tablenames
vector of tablenames to be used for corresponding files. By default, tables will be named using lowercase names from file basename with special characters replaced with underscores (for SQL compatibility).

...
additional arguments to streamable_table$read method.

Value

the database connection (invisibly)

Details

unark will read in a files in chunks and write them into a database. This is essential for processing large compressed tables which may be too large to read into memory before writing into a database. In general, increasing the lines parameter will result in a faster total transfer but require more free memory for working with these larger chunks.

If using readr-based streamable-table, you can suppress the progress bar by using options(readr.show_progress = FALSE) when reading in large files.

Examples

## Setup: create an archive.
library(dplyr)
dir <- tempdir()
db <- dbplyr::nycflights13_sqlite(tempdir())

## database -> .tsv.bz2 
ark(db, dir)

#> Warning: overwriting airlines.tsv.bz2

#> Exporting airlines in 50000 line chunks:

#> 	...Done! (in 0.00455308 secs)

#> Warning: overwriting airports.tsv.bz2

#> Exporting airports in 50000 line chunks:

#> 	...Done! (in 0.02540398 secs)

#> Warning: overwriting flights.tsv.bz2

#> Exporting flights in 50000 line chunks:

#> 	...Done! (in 18.91693 secs)

#> Warning: overwriting planes.tsv.bz2

#> Exporting planes in 50000 line chunks:

#> 	...Done! (in 0.04189301 secs)

#> Warning: overwriting weather.tsv.bz2

#> Exporting weather in 50000 line chunks:

#> 	...Done! (in 0.752681 secs)

## list all files in archive (full paths)
files <- list.files(dir, "[.]tsv\\.bz2$", full.names = TRUE)

## Read archived files into a new database (another sqlite in this case)
new_db <- src_sqlite(file.path(dir, "local.sqlite"), create=TRUE)
unark(files, new_db)

#> Importing airlines.tsv.bz2 in 50000 line chunks:

#> 	...Done! (in 0.01227999 secs)

#> Importing airports.tsv.bz2 in 50000 line chunks:

#> 	...Done! (in 0.02318192 secs)

#> Importing flights.tsv.bz2 in 50000 line chunks:

#> 	...Done! (in 8.445776 secs)

#> Importing planes.tsv.bz2 in 50000 line chunks:

#> 	...Done! (in 0.04314995 secs)

#> Importing weather.tsv.bz2 in 50000 line chunks:

#> 	...Done! (in 0.4591041 secs)

## Prove table is returned successfully.
tbl(new_db, "flights")

#> [90m# Source:   table<flights> [?? x 19][39m
#> [90m# Database: sqlite 3.22.0
#> #   [/var/folders/y8/0wn724zs10jd79_srhxvy49r0000gn/T//Rtmp7LUEtA/local.sqlite][39m
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m [3m[90m<int>[39m[23m    [3m[90m<int>[39m[23m          [3m[90m<int>[39m[23m     [3m[90m<int>[39m[23m    [3m[90m<int>[39m[23m          [3m[90m<int>[39m[23m
#> [90m 1[39m  [4m2[24m013     1     1      517            515         2      830            819
#> [90m 2[39m  [4m2[24m013     1     1      533            529         4      850            830
#> [90m 3[39m  [4m2[24m013     1     1      542            540         2      923            850
#> [90m 4[39m  [4m2[24m013     1     1      544            545        -[31m1[39m     [4m1[24m004           [4m1[24m022
#> [90m 5[39m  [4m2[24m013     1     1      554            600        -[31m6[39m      812            837
#> [90m 6[39m  [4m2[24m013     1     1      554            558        -[31m4[39m      740            728
#> [90m 7[39m  [4m2[24m013     1     1      555            600        -[31m5[39m      913            854
#> [90m 8[39m  [4m2[24m013     1     1      557            600        -[31m3[39m      709            723
#> [90m 9[39m  [4m2[24m013     1     1      557            600        -[31m3[39m      838            846
#> [90m10[39m  [4m2[24m013     1     1      558            600        -[31m2[39m      753            745
#> [90m# ... with more rows, and 11 more variables: arr_delay [3m[90m<int>[90m[23m, carrier [3m[90m<chr>[90m[23m,
#> #   flight [3m[90m<int>[90m[23m, tailnum [3m[90m<chr>[90m[23m, origin [3m[90m<chr>[90m[23m, dest [3m[90m<chr>[90m[23m, air_time [3m[90m<int>[90m[23m,
#> #   distance [3m[90m<int>[90m[23m, hour [3m[90m<int>[90m[23m, minute [3m[90m<int>[90m[23m, time_hour [3m[90m<dbl>[90m[23m[39m