Unarchive a list of compressed tsv files into a database

Unarchive a list of compressed tsv files into a database

unark(files, db_con, lines = 10000L, ...)

Arguments

files
vector of filenames to be read in. Must be tsv format, optionally compressed using bzip2, gzip, zip, or xz format at present.

db_con
a database src (src_dbi object from dplyr)

lines
number of lines to read in a chunk.

...
Arguments passed on to readr::read_tsv

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector). Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed. Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

quote

Single character used to quote strings.

col_names

Either TRUE, FALSE or a character vector of column names. If TRUE, the first row of the input will be used as the column names, and will not be included in the data frame. If FALSE, column names will be generated automatically: X1, X2, X3 etc. If col_names is a character vector, the values will be used as the names of the columns, and the first row of the input will be read into the first row of the output data frame. Missing (NA) column names will generate a warning, and be filled in with dummy names X1, X2 etc. Duplicate column names will generate a warning and be made unique with a numeric prefix.

col_types

One of NULL, a cols() specification, or a string. See vignette("column-types") for more details. If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself. If a column specification created by cols(), it must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only(). Alternatively, you can use a compact string representation where each character represents one column: c = character, i = integer, n = number, d = double, l = logical, D = date, T = date time, t = time, ? = guess, or _/- to skip the column.

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na

Character vector of strings to use for missing values. Set this option to character() to indicate no missing values.

quoted_na

Should missing values inside quotes be treated as missing values (the default) or strings.

comment

A string used to identify comments. Any text after the comment characters will be silently ignored.

trim_ws

Should leading and trailing whitespace be trimmed from each field before parsing it?

skip

Number of lines to skip before reading data.

n_max

Maximum number of records to read.

guess_max

Maximum number of records to use for guessing column types.

progress

Display a progress bar? By default it will only display in an interactive session and not while knitting a document. The display is updated every 50,000 values and will only display if estimated reading time is 5 seconds or more. The automatic progress bar can be disabled by setting option readr.show_progress to FALSE.

Value

the database connection (invisibly)

Details

unark will read in a files in chunks and write them into a database. This is essential for processing large compressed tables which may be too large to read into memory before writing into a database. In general, increasing the lines parameter will result in a faster total transfer but require more free memory for working with these larger chunks. #'

Examples

## Setup: create an archive.
library(dplyr)
dir <- tempdir()
db <- dbplyr::nycflights13_sqlite(tempdir())

## database -> .tsv.bz2 
ark(db, dir)

#> Warning: Closing open result set, pending rows

#> Exporting in 10000 line chunks:
#> airlines

#> Warning: Closing open result set, pending rows

#> ...Done! (in 0.01042795 secs)

#> Warning: Closing open result set, pending rows

#> Exporting in 10000 line chunks:
#> airports

#> Warning: Closing open result set, pending rows

#> ...Done! (in 0.03674603 secs)

#> Warning: Closing open result set, pending rows

#> Exporting in 10000 line chunks:
#> flights

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> ...Done! (in 12.29331 secs)

#> Warning: Closing open result set, pending rows

#> Exporting in 10000 line chunks:
#> planes

#> Warning: Closing open result set, pending rows

#> ...Done! (in 0.08438206 secs)

#> Warning: Closing open result set, pending rows

#> Exporting in 10000 line chunks:
#> weather

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> Warning: Closing open result set, pending rows

#> ...Done! (in 0.672559 secs)

## list all files in archive (full paths)
files <- list.files(dir, "[.]tsv\\.bz2$", full.names = TRUE)

## Read archived files into a new database (another sqlite in this case)
new_db <- src_sqlite(file.path(dir, "local.sqlite"), create=TRUE)
unark(files, new_db)

#> Importing in 10000 line chunks:
#> /var/folders/y8/0wn724zs10jd79_srhxvy49r0000gn/T//Rtmptt27wI/airlines.tsv.bz2

#> ...Done! (in 0.03802705 secs)

#> Importing in 10000 line chunks:
#> /var/folders/y8/0wn724zs10jd79_srhxvy49r0000gn/T//Rtmptt27wI/airports.tsv.bz2

#> ...Done! (in 0.0363431 secs)

#> Importing in 10000 line chunks:
#> /var/folders/y8/0wn724zs10jd79_srhxvy49r0000gn/T//Rtmptt27wI/flights.tsv.bz2

#> ...Done! (in 9.01262 secs)

#> Importing in 10000 line chunks:
#> /var/folders/y8/0wn724zs10jd79_srhxvy49r0000gn/T//Rtmptt27wI/planes.tsv.bz2

#> ...Done! (in 0.05085492 secs)

#> Importing in 10000 line chunks:
#> /var/folders/y8/0wn724zs10jd79_srhxvy49r0000gn/T//Rtmptt27wI/weather.tsv.bz2

#> ...Done! (in 0.3927159 secs)

## Prove table is returned successfully.
tbl(new_db, "flights")

#> # Source:   table<flights> [?? x 19]
#> # Database: sqlite 3.22.0
#> #   [/var/folders/y8/0wn724zs10jd79_srhxvy49r0000gn/T//Rtmptt27wI/local.sqlite]
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <int>    <int>          <int>
#>  1  2013     1     1      517            515         2      830            819
#>  2  2013     1     1      533            529         4      850            830
#>  3  2013     1     1      542            540         2      923            850
#>  4  2013     1     1      544            545        -1     1004           1022
#>  5  2013     1     1      554            600        -6      812            837
#>  6  2013     1     1      554            558        -4      740            728
#>  7  2013     1     1      555            600        -5      913            854
#>  8  2013     1     1      557            600        -3      709            723
#>  9  2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#> # ... with more rows, and 11 more variables: arr_delay <int>, carrier <chr>,
#> #   flight <int>, tailnum <chr>, origin <chr>, dest <chr>, air_time <int>,
#> #   distance <int>, hour <int>, minute <int>, time_hour <dbl>